feat: PlayerStatus에 플레이어 마나 변경 기능 구현
